cleanup and shuffle some code
